--- services/registration/src/api/mail/templates/index.html
@@ -1,53 +1,0 @@
-
-<!DOCTYPE html>
-<html>
-   <head>
-      <meta charset="UTF-8">
-      <style>
-         img { 
-               width: 100%;
-         }
-         body{
-            margin-top: 100px;
-            margin-bottom: 100px;
-            margin-right: 150px;
-            margin-left: 80px;
-            font-family: "HelveticaNeue-Light", "Helvetica Neue Light", "Helvetica Neue", Helvetica, Arial, "Lucida Grande", sans-serif;
-            font-size: large;
-         }
-         .confbody{
-            text-align: justify;
-            justify-content: space-around;
-            /* font-size: medium; */
-         }
-         .emoji{
-               width: 5px;
-               height: 5px;
-         }
-      </style>
-   </head>
-
-   <body>
-      <div id="banner">
-         <img src="../static/web-banner.png">
-      </div>
-      
-
-      <div id="confbody">
-         <p>
-               Hi {{ name }}! 
-         </p>
-         <p> 
-               Thanks for applying for CruzHacks 2019! There is so much we have planned for you in a couple of months and we can't wait to share what we have in store!
-         
-         </p>
-      </div>
-      <div id="signature"></div>
-         <p>
-            With Love<br>
-            CruzHacks 2019 Team
-         </p>
-         
-      </div>
-   </body>
-</html>
